dev-requirements.txt
coverage==7.2.1
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2-binary
drf_yasg
django-environ==0.9.0
Pillow
sorl-thumbnail
django-cleanup
sorl-thumbnail-serializer-field==0.2.1
django-cors-headers
djangochannelsrestframework
channels
channels-redis
django-filter
dj-rest-auth